Student Dash
Made significant changes to calendar functionality

Calendar is now populated with days and the alignment with the days of
the week issue is resolved.

Calendar now displays and notes days from previous and future months
existing on the same calendar.

Calendar now is traversable and recognizes the correct date to set to be
active